add head of family

+++ database/migrations/2025_09_12_095025_create_head_of_families_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->uuid('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->string('profile_picture');
-            $table->integer('identity_number');
+            $table->bigInteger('identity_number');
             $table->enum('gender', ['male','female']);
             $table->date('date_of_birth');
             $table->string('phone_number');
